# openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/users/query":{"get":{"operationId":"UsersController_query","parameters":[{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"age","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/items":{"get":{"operationId":"ItemsController_findAll","parameters":[],"responses":{"200":{"description":"The found records","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponseDto"}}}}}}}},"/items/{id}":{"get":{"operationId":"ItemsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemResponseDto"}}}}}}}},"info":{"title":"Swagger Test","description":"The Swagger Test API description","version":"1.0","contact":{}},"tags":[{"name":"swagger","description":""}],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"id":{"type":"string","example":"uuid"},"name":{"type":"string","example":"Shinya"},"age":{"type":"number","example":29}},"required":["id","name","age"]},"ItemResponseDto":{"type":"object","properties":{"id":{"type":"string","readOnly":true,"example":"uuid"},"name":{"type":"string","readOnly":true,"example":"Mac book Air 2020"},"description":{"type":"string","readOnly":true,"example":"この商品はMacBookの2020年モデルです"},"price":{"type":"number","readOnly":true,"example":220000},"stock":{"type":"number","readOnly":true,"example":3},"imageUrl":{"type":"string","readOnly":true,"example":"https://www.google.com/.png"},"createdAt":{"type":"string","readOnly":true,"example":"2024-03-09T16:39:13.754Z"},"updatedAt":{"type":"string","readOnly":true,"example":"2024-03-09T16:39:13.754Z"},"stockMessage":{"type":"string","readOnly":true,"example":"在庫わずかです。"}},"required":["id","name","description","price","stock","imageUrl","createdAt","updatedAt","stockMessage"]}}}}